+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
+- Simplify output of `schemalint verify` to improve appearance in GitHub actions log.
 - Add check that `additionalProperties` is disabled on all objects.
 
 ## [0.6.0] - 2023-02-02