+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2023-02-06
+
 - Add check that every array specifies the schema of its items with the `items` keyword.
 - Simplify output of `schemalint verify` to improve appearance in GitHub actions log.
 - Add check that `additionalProperties` is disabled on all objects.
@@ -52,7 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added first basic linting.
 
-[Unreleased]: https://github.com/giantswarm/schemalint/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/giantswarm/schemalint/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/giantswarm/schemalint/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/giantswarm/schemalint/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/giantswarm/schemalint/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/giantswarm/schemalint/compare/v0.3.0...v0.4.0